Fix the incorrect fetch partial parsing
Fix #131

The original implementation interpreted the partial fetch arguments as
a start and end octet, but the RFC states they are a start octet and a
length. This also causes incorrect parsing failures because there was a
check that the first number was less than the second number, which does
not make sense with corrected assumptions.